+++ crates/test/src/lib.rs
@@ -136,3 +136,5 @@ pub use __rt::criterion::Criterion;
 
 // web_time Instant
 pub use __rt::web_time::Instant;
+
+pub use __rt::browser::{screenshot, Screenshot, ScreenshotError};